추가 디자인 적용
