+++ src/app/keywords/KeywordRankList.tsx
@@ -8,6 +8,7 @@ import SvgIcon from '@/components/icon/SvgIcon';
 import { useRouter } from 'next/navigation';
 import { ROUTE_PATH } from '@/constants/route';
 import { map } from 'es-toolkit/compat';
+import { useKeywordRankFilter } from '@/hooks/useKeywordRankFilter';
 
 interface Props {
   title: string;
@@ -20,6 +21,7 @@ export default function KeywordRankList({ title, list }: Props) {
   const router = useRouter();
   const [activeRank, setActiveRank] = useState<number>(0);
   const [isHovered, setIsHovered] = useState(false);
+  const { filterPeriod } = useKeywordRankFilter();
   const intervalRef = useRef<NodeJS.Timeout | null>(null);
   const intervalTime = 1500;
   useEffect(() => {
@@ -48,7 +50,7 @@ export default function KeywordRankList({ title, list }: Props) {
         clearInterval(intervalRef.current);
       }
     };
-  }, [isHovered, list, intervalTime, intervalRef.current]);
+  }, [isHovered, list, intervalTime, intervalRef]);
 
   return (
     <div className="flex flex-col w-full h-full">
@@ -74,9 +76,23 @@ export default function KeywordRankList({ title, list }: Props) {
           </div>
         ))}
         {list.length > 0 && (
-          <div className="h-[67px] flex flex-col items-end text-greyscale-60 mt-3 text-[10px]">
-            <span>서비스에서 제공하는 키워드는 특정 기준에 따라 선정된 것으로</span>
-            <span>실제 실시간 키워드와 다를 수 있습니다.</span>
+          <div className="px-2 py-3 bg-greyscale-80 flex flex-col mt-3 text-[10px] rounded-[12px]">
+            <div className="flex text-greyscale-30">
+              <span className="mr-2">•</span>
+              <div className="flex flex-col">
+                <span>서비스에서 제공하는 키워드는 특정 기준에 따라 선정된 것으로,</span>
+                <span>실제 실시간 키워드와 다를 수 있습니다.</span>
+              </div>
+            </div>
+            {filterPeriod !== 'realtime' && (
+              <div className="flex text-primary-80 mt-3">
+                <span className="mr-2">•</span>
+                <div className="flex flex-col">
+                  <span>기간 필터 적용시, 키워드 별 채팅하기를 이용할 수 없습니다.</span>
+                  <span>채팅하기 서비스를 이용하시려면 실시간 키워드를 활용해 주세요.</span>
+                </div>
+              </div>
+            )}
           </div>
         )}
       </ScrollArea>
